+++ gantry/commands/run.py
@@ -219,9 +219,9 @@ from .main import CLICK_COMMAND_DEFAULTS, config, main, new_optgroup
 @optgroup.option(
     "--gh-token-secret",
     type=str,
-    help="""The name of the Beaker secret that contains your GitHub token.""",
-    default=constants.GITHUB_TOKEN_SECRET,
-    show_default=True,
+    help=f"""The name of the Beaker secret that contains your GitHub token.
+    {config.get_help_string_for_default('gh_token_secret', constants.GITHUB_TOKEN_SECRET)}""",
+    default=config.gh_token_secret or constants.GITHUB_TOKEN_SECRET,
 )
 @new_optgroup("Outputs")
 @optgroup.option(