+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `--text` option to `gantry list` command for filtering by name or description.
 - Added `client` parameter to `api.update_workload_description()` for providing an existing Beaker client,
   which avoids creating one each time the function is called.
+- Added support for configuring the GitHub token secret name in a `pyproject.toml` as the field `[tool.gantry.gh_token_secret]`.
 
 ### Changed